Upgrade n8n 09 14 2023

+++ packages/editor-ui/src/views/VariablesView.vue
@@ -12,7 +12,7 @@ import { useI18n, useTelemetry, useToast, useMessage } from '@/composables';
 import ResourcesListLayout from '@/components/layouts/ResourcesListLayout.vue';
 import VariablesRow from '@/components/VariablesRow.vue';
 
-import { EnterpriseEditionFeature } from '@/constants';
+import { EnterpriseEditionFeature, MODAL_CONFIRM } from '@/constants';
 import type {
 	DatatableColumn,
 	EnvironmentVariable,
@@ -26,7 +26,7 @@ const environmentsStore = useEnvironmentsStore();
 const usersStore = useUsersStore();
 const uiStore = useUIStore();
 const telemetry = useTelemetry();
-const { i18n } = useI18n();
+const i18n = useI18n();
 const message = useMessage();
 const sourceControlStore = useSourceControlStore();
 let sourceControlStoreUnsubscribe = () => {};
@@ -183,7 +183,7 @@ function cancelEditing(data: EnvironmentVariable | TemporaryEnvironmentVariable)
 
 async function deleteVariable(data: EnvironmentVariable) {
 	try {
-		await message.confirm(
+		const confirmed = await message.confirm(
 			i18n.baseText('variables.modals.deleteConfirm.message', { interpolate: { name: data.key } }),
 			i18n.baseText('variables.modals.deleteConfirm.title'),
 			{
@@ -191,11 +191,11 @@ async function deleteVariable(data: EnvironmentVariable) {
 				cancelButtonText: i18n.baseText('variables.modals.deleteConfirm.cancelButton'),
 			},
 		);
-	} catch (e) {
-		return;
-	}
 
-	try {
+		if (confirmed !== MODAL_CONFIRM) {
+			return;
+		}
+
 		await environmentsStore.deleteVariable(data);
 		allVariables.value = allVariables.value.filter((variable) => variable.id !== data.id);
 	} catch (error) {
@@ -228,6 +228,7 @@ onBeforeUnmount(() => {
 
 <template>
 	<ResourcesListLayout
+		class="variables-view"
 		ref="layoutRef"
 		resource-key="variables"
 		:disabled="!isFeatureEnabled"
@@ -275,11 +276,12 @@ onBeforeUnmount(() => {
 				"
 				:buttonText="$locale.baseText(contextBasedTranslationKeys.variables.unavailable.button)"
 				buttonType="secondary"
-				@click="goToUpgrade"
+				@click:button="goToUpgrade"
 			/>
 		</template>
-		<template v-if="!isFeatureEnabled" #empty>
+		<template v-if="!isFeatureEnabled || (isFeatureEnabled && !canCreateVariables)" #empty>
 			<n8n-action-box
+				v-if="!isFeatureEnabled"
 				data-test-id="unavailable-resources-list"
 				emoji="👋"
 				:heading="$locale.baseText(contextBasedTranslationKeys.variables.unavailable.title)"
@@ -288,6 +290,18 @@ onBeforeUnmount(() => {
 				"
 				:buttonText="$locale.baseText(contextBasedTranslationKeys.variables.unavailable.button)"
 				buttonType="secondary"
+				@click:button="goToUpgrade"
+			/>
+			<n8n-action-box
+				v-else-if="!canCreateVariables"
+				data-test-id="cannot-create-variables"
+				emoji="👋"
+				:heading="
+					$locale.baseText('variables.empty.notAllowedToCreate.heading', {
+						interpolate: { name: usersStore.currentUser.firstName },
+					})
+				"
+				:description="$locale.baseText('variables.empty.notAllowedToCreate.description')"
 				@click="goToUpgrade"
 			/>
 		</template>
@@ -318,43 +332,45 @@ onBeforeUnmount(() => {
 <style lang="scss" scoped>
 @use 'n8n-design-system/css/common/var.scss';
 
-:deep(.datatable) {
-	table {
-		table-layout: fixed;
-	}
-
-	th,
-	td {
-		width: 25%;
-
-		@media screen and (max-width: var.$md) {
-			width: 33.33%;
+.variables-view {
+	:deep(.datatable) {
+		table {
+			table-layout: fixed;
 		}
 
-		&.variables-value-column,
-		&.variables-key-column,
-		&.variables-usage-column {
-			> div {
-				width: 100%;
+		th,
+		td {
+			width: 25%;
 
-				> span {
-					max-width: 100%;
-					overflow: hidden;
-					text-overflow: ellipsis;
-					white-space: nowrap;
-					height: 18px;
-				}
+			@media screen and (max-width: var.$md) {
+				width: 33.33%;
+			}
 
+			&.variables-value-column,
+			&.variables-key-column,
+			&.variables-usage-column {
 				> div {
 					width: 100%;
+
+					> span {
+						max-width: 100%;
+						overflow: hidden;
+						text-overflow: ellipsis;
+						white-space: nowrap;
+						height: 18px;
+					}
+
+					> div {
+						width: 100%;
+					}
 				}
 			}
 		}
-	}
 
-	.variables-usage-column {
-		@media screen and (max-width: var.$md) {
-			display: none;
+		.variables-usage-column {
+			@media screen and (max-width: var.$md) {
+				display: none;
+			}
 		}
 	}
 }
